pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-sqlalchemy>=3.1.1",
    "flask>=3.1.0",
    "openai>=1.65.3",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "psycopg2-binary>=2.9.10",
    "email-validator>=2.2.0",
    "gunicorn>=23.0.0",
    "sqlalchemy>=2.0.38",
    "pillow>=11.1.0",
    "flask-cors>=5.0.1",
    "ollama>=0.5.1",
]
